+++ httpServer.cpp
@@ -268,7 +268,7 @@ bool HttpServer::parseParams(ParserUtils &pu, ClientContext &ctx)
 
             try
             {
-                ctx.gain = std::stoi(pu.token().toString()); // throw exc
+                ctx.gain = std::stoul(pu.token().toString()); // throw exc
                 ctx.gainFound = true;
             }
             catch (const std::exception &exc)
